requirements.txt
python-dotenv==1.0.1
langchain==0.1.10
langchain-openai==0.0.8
huggingface-hub==0.21.3
wikipedia==1.4.0
chromadb==0.4.24
tiktoken==0.6.0
